Ajout de fonctionnalites
- ajout des dechets aleatoires
- Paiement des visiteurs a chaque minute
- Deduction d'argent pour chaque dechet
- refactor de code repetitif

+++ TP2 Zoo/Personne/Heros.cs
@@ -7,12 +7,12 @@ using System.Windows.Forms;
 
 namespace TP2_Zoo.Personne {
     public class Heros : Personne {
-        public int Argent { get; set; }
+        public double Argent { get; set; }
         public int Touche { get; set; }
 
         public Heros() {
             this.Position = new int[] { 1, 5 };
-            Argent = 100;
+            Argent = 100.0;
 
             Sprites.Add(Properties.Resources.bas1);     // 0
             Sprites.Add(Properties.Resources.droite1);  // 1
